move OnDiskOrdinalsMap to its own class, and move write() code into it
