+++ src/java/org/apache/cassandra/index/sai/disk/hnsw/CassandraOnHeapHnsw.java
@@ -21,7 +21,6 @@ package org.apache.cassandra.index.sai.disk.hnsw;
 import java.io.IOException;
 import java.nio.ByteBuffer;
 import java.util.Collection;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.PriorityQueue;
@@ -30,8 +29,6 @@ import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.locks.ReadWriteLock;
 import java.util.concurrent.locks.ReentrantReadWriteLock;
-
-import com.google.common.base.Preconditions;
 
 import org.apache.cassandra.db.marshal.ByteBufferAccessor;
 import org.apache.cassandra.db.marshal.ValueAccessor;
@@ -61,12 +58,12 @@ import org.apache.lucene.util.hnsw.RandomAccessVectorValues;
 @SuppressWarnings("com.google.common.annotations.Beta")
 public class CassandraOnHeapHnsw
 {
-    private final ByteBufferVectorValues vectorValues;
+    final ByteBufferVectorValues vectorValues;
     private final ConcurrentHnswGraphBuilder<float[]> builder;
     private final AtomicInteger cachedDimensions = new AtomicInteger();
     private final TypeSerializer<float[]> serializer;
     private final VectorSimilarityFunction similarityFunction;
-    private final Map<ByteBuffer, VectorPostings> postingsMap;
+    final Map<ByteBuffer, VectorPostings> postingsMap;
     private final AtomicInteger nextOrdinal = new AtomicInteger();
 
     // FIXME this is disgusting and possibly unnecessary
@@ -173,40 +170,9 @@ public class CassandraOnHeapHnsw
         return vectorValues.size();
     }
 
-    private int rowCount()
+    int rowCount()
     {
         return postingsMap.values().stream().mapToInt(p -> p.keys.size()).sum();
-    }
-
-    private void writeOrdinalToRowMapping(File file, Map<PrimaryKey, Integer> keyToRowId) throws IOException
-    {
-        Preconditions.checkState(keyToRowId.size() == rowCount(),
-                                 "Expected %s rows, but found %s", keyToRowId.size(), rowCount());
-        Preconditions.checkState(postingsMap.size() == vectorValues.size(),
-                                 "Postings entries %s do not match vectors entries %s", postingsMap.size(), vectorValues.size());
-        try (var iow = IndexFileUtils.instance.openOutput(file)) {
-            var out = iow.asSequentialWriter();
-            // total number of vectors
-            out.writeInt(vectorValues.size());
-
-            // Write the offsets of the postings for each ordinal
-            long offset = 4L + 8L * vectorValues.size();
-            for (var i = 0; i < vectorValues.size(); i++) {
-                // (ordinal is implied; don't need to write it)
-                out.writeLong(offset);
-                var postings = postingsMap.get(vectorValues.bufferValue(i));
-                offset += 4 + (postings.keys.size() * 4L); // 4 bytes for size and 4 bytes for each integer in the list
-            }
-
-            // Write postings lists
-            for (var i = 0; i < vectorValues.size(); i++) {
-                var postings = postingsMap.get(vectorValues.bufferValue(i));
-                out.writeInt(postings.keys.size());
-                for (var key : postings.keys) {
-                    out.writeInt(keyToRowId.get(key));
-                }
-            }
-        }
     }
 
     private void writeGraph(File file) throws IOException
@@ -235,7 +201,7 @@ public class CassandraOnHeapHnsw
         try
         {
             writeVectors(descriptor.fileFor(IndexComponent.VECTOR, context));
-            writeOrdinalToRowMapping(descriptor.fileFor(IndexComponent.POSTING_LISTS, context), keyToRowId);
+            OnDiskOrdinalsMap.writeOrdinalToRowMapping(this, descriptor.fileFor(IndexComponent.POSTING_LISTS, context), keyToRowId);
             writeGraph(descriptor.fileFor(IndexComponent.TERMS_DATA, context));
         }
         finally
@@ -244,7 +210,7 @@ public class CassandraOnHeapHnsw
         }
     }
 
-    private static class VectorPostings
+    static class VectorPostings
     {
         public final int ordinal;
         public final List<PrimaryKey> keys;
@@ -262,7 +228,7 @@ public class CassandraOnHeapHnsw
         }
     }
 
-    private class ByteBufferVectorValues implements RandomAccessVectorValues<float[]>
+    class ByteBufferVectorValues implements RandomAccessVectorValues<float[]>
     {
         private final Map<Integer, ByteBuffer> values = new ConcurrentSkipListMap<>();
